project/app/pyproject.toml
[project]
name = "pydantic-doc-agent-app"
version = "0.1.0"
description = "Day 6 homework: Streamlit + CLI for Pydantic docs agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minsearch>=0.0.10",
    "openai>=2.30.0",
    "pydantic-ai>=1.73.0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "requests>=2.33.0",
    "streamlit>=1.40.0",
]
